use a clean Makefile system to build/deploy site
This way, we get dependency/smart update for free. We don't need to
rebuild all the pages everytime a new one is added. The author, date of
creation and date of revision are extracted from git information. The
index is generated automatically, so adding a recipe is literally just
adding a markdown file in src/.

The deployment system uses rsync to push only newer data to the server.

There is a tag system that you can add at the end of each recipe file.
Tag pages are generated statically that contain all recipes with a given
tag.

It introduces more HTML template files, though.

To build website, type
make build

To deploy it,
make deploy

To clean everything,
make clean