Update carbonara.md
That bowl mix was a nightmare for my eyes, unbelievable that such a heresy can be achieved. I included direction 4 as contemporary to 2 and 3: you'll be more efficient this way, since beating the eggs takes about five minutes. Also I never understood why everybody says it's easy, but I guess it is..